Reduce Garbage Collection form Stirng usage
While processing a large number of features, we noticed
Garbage Collection cycles of > 10 seconds. This was due to the
large number of created strings used to encode and decode the
feature names.

This commit sets the feature names to be resuable String builders
and exposes methods to setName as int, double, String, float and long.
The murmur hash function now also accepts ByteBuffers that are reusable
and are created when the feature is initially created and are reused
and reset along with the string builder objects and the encompassing
reusable CharBuffer as well.

The StringBuilder is kept there to be able to easily debug, althoug it
adds a bit more memory, it is reusable memory and very predictable in
terms of added Garbage for the JVM to take care of.